use float for dist_t and cost_t

+++ sound_law/rl/common.h
@@ -15,8 +15,8 @@ using namespace std;
 
 // Define basic integer types for different purporses.
 using abc_t = uint16_t;    // for alphabet
-using cost_t = uint8_t;    // for cost
-using dist_t = uint16_t;   // for distance
+using cost_t = float;      // for cost
+using dist_t = float;      // for distance
 using visit_t = int32_t;   // for visit/action counts -- due to virtual games, this could be negative.
 using action_t = uint32_t; // for actions
 using node_t = uint64_t;   // for node id
